Move UndefMask and Relocations into `allocation.rs`
